SFTP API Documentation

Base.Filesystem.basename — Function

basename(uri::SFTP.URI, path::AbstractString=".") -> String
basename(sftp::SFTP.Client, path::AbstractString=".") -> String

Get the file name or current folder name of a path. The path can be absolute or relative to the uri or current directory of the sftp server given. If no path is given, the current path from the uri or sftp server is taken.

Note

In contrast to Julias basename, trailing slashes in paths are ignored and the last non-empty part is returned, except for the root, where the basename is empty.

Base.Filesystem.cd — Method

cd(sftp::SFTP.Client, dir::AbstractString)

Change to dir in the uri of the sftp client.

Base.Filesystem.filemode — Function

filemode(sftp::SFTP.Client, path::AbstractString = ".") -> UInt
filemode(st::SFTP.StatStruct) -> UInt

Return the filemode of the SFTP.StatStruct. A convenience method exists to directly check the path on the sftp server. However, if several path objects in the same folder are analysed, it is much more performant to use statscan once and then analyse each SFTP.StatStruct.

Base.Filesystem.isdir — Function

isdir(sftp::SFTP.Client, path::AbstractString = ".") -> Bool
isdir(st::SFTP.StatStruct) -> Bool

Analyse the SFTP.StatStruct and return true for a directory, false otherwise. A convenience method exists to directly check the path on the sftp server. However, if several path objects in the same folder are analysed, it is much more performant to use statscan once and then analyse each SFTP.StatStruct.

Base.Filesystem.isfile — Function

isfile(sftp::SFTP.Client, path::AbstractString = ".") -> Bool
isfile(st::SFTP.StatStruct) -> Bool

Analyse the SFTP.StatStruct and return true for a file, false otherwise. A convenience method exists to directly check the path on the sftp server. However, if several path objects in the same folder are analysed, it is much more performant to use statscan once and then analyse each SFTP.StatStruct.

Base.Filesystem.islink — Function

islink(sftp::SFTP.Client, path::AbstractString = ".") -> Bool
islink(st::SFTP.StatStruct) -> Bool

Analyse the SFTP.StatStruct and return true for a symbolic link, false otherwise. A convenience method exists to directly check the path on the sftp server. However, if several path objects in the same folder are analysed, it is much more performant to use statscan once and then analyse each SFTP.StatStruct.

Base.Filesystem.ispath — Function

ispath(sftp::Client, path::AbstractString = ".") -> Bool

Return true, if a path exists on the sftp server, i.e. is a file, folder or link. Otherwise, reture false.

Base.Filesystem.joinpath — Function

joinpath(sftp::SFTP.Client, path::AbstractString...; kwargs...) -> URI
joinpath(sftp::URI, path::AbstractString...; kwargs...) -> URI

Join any path with the uri of the sftp client or the uri directly and return an URI with the updated path.

Note

The uri field of the sftp client remains unaffected by joinpath. Use sftp.uri = joinpath(sftp, "new/path") to update the URI on the sftp client.

Base.Filesystem.mkdir — Method

mkdir(sftp::SFTP.Client, dir::AbstractString) -> String

Create a new dir on the sftp server and return the name of the created directory. Although a path can be given as dir, dir can only be created in an existing directory, i.e. the path up to the basename of dir must exist. Otherwise, and in case of already existing folders, an error is thrown.

Base.Filesystem.mkpath — Method

mkpath(sftp::SFTP.Client, path::AbstractString) -> String

Create a path on the sftp client and return path as String on success.

Base.Filesystem.mv — Method

mv(
    sftp::SFTP.Client,
    src::AbstractString,
    dst::AbstractString;
    force::Bool=false
)

Move src to dst in the uri of the sftp client. The parent folder dst is moved to must exist. The src is overwritten without warning, if force is set to true.

Base.Filesystem.pwd — Function

pwd(sftp::SFTP.Client) -> String
pwd(uri::SFTP.URI) -> String

Return the current URI path of the SFTP Client or an URI struct. If a SFTP.Client is given, pwd checks whether the path is valid and throws an IOError otherwise. For URI there are no validity checks.

Base.Filesystem.readdir — Function

readdir(
    sftp::SFTP.Client,
    path::AbstractString = ".";
    join::Bool = false,
    sort::Bool = true,
    check_path::Bool = false
) -> Vector{String}

Read the current directory on the sftp client and return a vector of strings with the file names just like Julia's readdir. If join is set to true, the list of file names include the absolute path. Sorting of file names can be switched off with the sort flag to optimise performance. Depending on the server settings, readdir may return an empty vector for non-existant paths. To ensure an error is thrown for non-existant paths, set check_path to true.

Note

Setting check_path to true can drastically reduce the performance for large existing folders. If you know the folder structure, you should avoid setting this flag.

Base.Filesystem.rm — Method

rm(sftp::Client, path::AbstractString; recursive::Bool=false, force::Bool=false)

Remove (delete) the path on the sftp client. Set the recursive flag to remove folders recursively. Suppress errors by setting force to true.

Warning

Recursive deletions can be very slow for large folders.

Base.Filesystem.splitdir — Function

splitdir(uri::SFTP.URI, path::AbstractString=".") -> Tuple{URI,String}
splitdir(sftp::SFTP.Client, path::AbstractString=".") -> Tuple{URI,String}

Join the path with the path of the URI in sftp (or itself, if only a URI is given) and then split it into the directory name and base name. Return a Tuple of URI with the split path and a String with the base name.

Base.Filesystem.walkdir — Function

walkdir(
    sftp::SFTP.Client,
    root::AbstractString=".";
    topdown::Bool=true,
    follow_symlinks::Bool=false,
    skip_restricted_access::Bool=true,
    sort::Bool=true
) -> Channel{Tuple{String,Vector{String},Vector{String}}}

Return an iterator that walks the directory tree of the given root on the sftp client. If the root is omitted, the current URI path of the sftp client is used. The iterator returns a tuple containing (rootpath, dirs, files). The iterator starts at the root unless topdown is set to false.

If follow_symlinks is set to true, the sources of symlinks are listed rather than the symlink itself as a file. If sort is set to true, the files and directories are listed alphabetically. If a remote folder has restricted access, these directories are skipped with an info output on the terminal unless skip_restricted_access is set to false, in which case an Downloads.RequestError is thrown.

Examples

sftp = SFTP.Client("sftp://test.rebex.net/pub/example/", "demo", "password")
for (root, dirs, files) in walkdir(sftp, "/")
    println("Directories in $root")
    for dir in dirs
        println(joinpath(root, dir)) # path to directories
    end
    println("Files in $root")
    for file in files
        println(joinpath(root, file)) # path to files
    end
end

Base.download — Function

download(
    sftp::Client,
    src::AbstractString = ".",
    dst::AbstractString = ".";
    merge::Bool = false,
    force::Union{Nothing,Bool} = nothing,
    ignore_hidden::Bool = false,
    hide_identifier::Union{Char,AbstractString} = '.'
) -> String

Download src from the sftp server to dst on the local system; src can be a file or folder. Folders are downloaded recursively. dst must be an existing folder on the local system, otherwise an IOError is thrown. src may include an absolute or relative path on the sftp server, which is ignored on the local system. dst can be an absolute or relative path on the local system. The function returns dst as String.

kwargs

merge::Bool=false: download into existing folders, when true or throw an IOError

force::Union{Nothing,Bool}=nothing: Handle conflicts of existing files/folders

nothing (default): throw IOError for conflicts

true: overwrite existing files/folders

false: skip existing files/folders

if merge = true, force counts only for files

ignore_hidden::Bool=false: ignore hidden files and folders

hide_identifier::Union{Char,AbstractString}='.': start sequence of hidden files and folders

Example

sftp = SFTP.Client("sftp://test.rebex.net/pub/example/", "demo", "password")
files=readdir(sftp)
download_dir="/tmp"
download.(sftp, files, download_dir)

Alternatively:

sftp = SFTP.Client("sftp://test.rebex.net/pub/example/", "demo", "password")
donwload(sftp) # downloads current folder on server to current directory on local system

Base.download — Method

download(fcn::Function, sftp::Client, src::AbstractString)

Download src from the sftp server and use fcn to retrieve the data from src.srcmay include an absolute or relative path to the file on thesftp` server. Only temporary files are created on the local system and directly deleted after data reading.

.. info: fcn must have one AbstractString parameter, which is used to pass the temporary file path to any function that reads the contents. The function must return the contents of the file in any desired format, e.g. Matrix, DataFrame or array.

Examples

using CSV, DataFrames
# Define functions to process the file data
fcsv(path::AbstractString)::Matrix{Int} = CSV.read(path, CSV.Tables.matrix)
fread(path::AbstractString)::Vector{String} = readlines(path)
# Download data to variable
matrix = download(fcsv, sftp, "data/matrix.csv", _test=true)
array = download(fread, sftp, "data/matrix.csv", _test=true)

Base.isequal — Method

isequal(a::SFTP.StatStruct, b::SFTP.StatStruct) -> Bool

Compares equality between the description (desc fields) of two SFTP.StatStruct objects and returns true for equality, otherwise false.

Base.isless — Method

isless(a::SFTP.StatStruct, b::SFTP.StatStruct) -> Bool

Compares the descriptions (desc fields) of two SFTP.StatStruct objects and returns true, if a is lower than b, otherwise false.

Base.stat — Function

stat(sftp::SFTP.Client, path::AbstractString=".") -> SFTP.StatStruct

Return the stat data for path on the sftp server.

Note

This returns only stat data for one object, but stat data for all objects in the same folder is obtained internally. If you need stat data for more than object in the same folder, use statscan for better performance and reduced connections to the server.

SFTP.analyse_path — Method

analyse_path(sftp::SFTP.Client, root::AbstractString) -> Bool

Return, whether the root on the sftp server is a directory or a link pointing to a directory and the path of the directory.

SFTP.change_uripath — Method

change_uripath(sftp::Client, path::AbstractString...) -> URI
change_uripath(uri::URI, path::AbstractString; trailing_slash::Union{Bool,Nothing}=nothing) -> URI

Return an updated uri struct with the given path. When an sftp client is passed, a trailing slash will be added for directories and omitted otherwise. If a uri struct is passed, a trailing_slash is added or omitted, when the flag is true/false, or left unchanged, if trailing_slash is nothing.

.. warning:: Determining directories for the method using the sftp client can be slow for large folders and is not recommended unless absolutely needed.

SFTP.check_and_create_fingerprint — Function

check_and_create_fingerprint(
    host::AbstractString,
    known_hosts_file=joinpath(homedir(), ".ssh", "known_hosts")
)

)

Check for host in known_hosts_file.

SFTP.create_fingerprint — Method

create_fingerprint(
    host::AbstractString,
    known_hosts::AbstractString,
    content::Vector{String}
)

Create a new entry in known_hosts for host adding it in front of the existing rows.

SFTP.download_file — Function

download_file(
    sftp::Client,
    src::URI,
    dst::AbstractString,
    [files::Vector{<:AbstractString},]
    conflicts::Vector{<:AbstractString},
    force::Union{Nothing,Bool}
)

Download src from the sftp server to dst on the local system; src can either be a file or a path to a folder on the sftp server. In the latter case, files must be defined as a vector of strings and potential conflicts on the server given as additional vector of strings. The force flag determines, whether existing files are overwritten (true), skipped (false) or throw an error (nothing).

SFTP.findbase — Method

findbase(stats::Vector{SFTP.StatStruct}, base::AbstractString, path::AbstractString) -> Int

Return the index of base in stats or throw an IOError, if base is not found.

SFTP.ftp_command — Method

ftp_command(sftp::SFTP.Client, command::String)

Execute the command on the sftp server.

SFTP.isfalse — Method

isfalse(x)::Bool

Return true if x is a Bool and false, otherwise false.

SFTP.istrue — Method

istrue(x)::Bool

Return true if x is a Bool and true, otherwise false.

SFTP.joinrootpath — Method

joinrootpath(root::AbstractString, parts::AbstractString...) -> String

Join all path parts to the root path and return as String. The parts are joined to the root even if the yield an absolute path.

Note

This is an internal version for testing mocked file uploads. The root is expected to be a tempdir.

SFTP.mkfolder — Function

mkfolder(
    [sftp::Client,]
    path::AbstractString,
    dirs::Vector{<:AbstractString},
    conflicts::Vector{<:AbstractString},
    merge::Bool,
    force::Union{Nothing, Bool},
    [ignore_hidden::Bool,
    hide_identifier::Union{AbstractString,Char}]
)

Create the given path either on the sftp server or the local system, depending on whether the sftp client was passed as first argument. Ignore hidden folders starting with the hide_identifier, if ignore_hidden is true. For the local system, hidden folders are filtered by walkdir in the download function.

Only non-existing dirs from the source system are created on the destination. Existing path conflicts on the destination system are handled according to the flag settings of merge and force. Only throw an error for existing path conflicts, if merge is set to false and force is set to nothing. Delete conflicts, if force is true and merge is false, and ignore conflicts, if force is false or merge is true.

SFTP.parse_date — Method

parse_date(month::AbstractString, day::AbstractString, year_or_time::AbstractString) -> Float64

From the abbreviated month name, the day and the year_or_time all given as String, return a unix timestamp.

SFTP.parse_mode — Method

parse_mode(s::AbstractString) -> UInt

From the AbstractString s, parse the file mode octal number and return as UInt.

SFTP.reset_easy_hook — Method

reset_easy_hook(sftp::SFTP.Client) -> Nothing

Reset curl easy options to standard as defined by the sftp client.

SFTP.set_stdopt — Method

set_stdopt(sftp::SFTP.Client, easy::Easy) -> Nothing

Set defaults for a number of curl easy options as defined by the sftp client.

SFTP.set_uri — Method

set_uri(uri::URI) -> URI

Return a URI struct from the given uri.

SFTP.statscan — Function

statscan(
    sftp::SFTP.Client,
    path::AbstractString=".";
    sort::Bool=true,
    show_cwd_and_parent::Bool=false
) -> Vector{SFTP.StatStruct}

Like stat, but returns a Vector of SFTP.StatStruct with filesystem stats for all objects in the given path.

** This should be preferred over stat for performance reasons. **

Note

You can only run this on directories.

By default, the SFTP.StatStruct vector is sorted by the descriptions (desc fields). For large folder contents, sort can be set to false to increase performance, if the output order is irrelevant. If show_cwd_and_parent is set to true, the SFTP.StatStruct vector includes entries for "." and ".." on position 1 and 2, respectively.

SFTP.str2number — Method

str2number(s::AbstractString) -> Int64

Parse the file owner symbols in the string s to the corresponding ownership number.

SFTP.symlink_target! — Method

symlink_target!(
    sftp::SFTP.Client,
    stats::SFTP.StatStruct,
    root::AbstractString,
    dirs::Vector{String},
    files::Vector{String},
    follow_symlinks::Bool
)

Analyse the symbolic link on the sftp server from its stats and add it to the respective dirs or files list. The root path is needed to get updated stats of the symlink. Save the source of the symlink, if follow_symlinks is set to true, otherwise save symlinks as files.

SFTP.unescape_joinpath — Method

unescape_joinpath(sftp::SFTP.Client, path::AbstractString) -> String

Join the path with the URI path in sftp and return the unescaped path. Note, this function should not use URL:s since CURL:s api need spaces

SFTP.upload — Function

upload(
    sftp::Client,
    src::AbstractString=".",
    dst::AbstractString=".";
    merge::Bool=false,
    force::Bool=false,
    ignore_hidden::Bool=false,
    hide_identifier::Union{Char,AbstractString}='.'
) -> String

Upload (put) src to dst on the server; src can be a file or folder. Folders are uploaded recursively. dst must be an existing folder on the server, otherwise an IOError is thrown. src may include an absolute or relative path on the local system, which is ignored on the server. dst can be an absolute path or a path relative to the current uri path of the sftp server. The function returns dst as String.

If merge is set to true, the content of src is merged into any existing dst folder. If force is set to true, any existing path at dst on the sftp server is overwritten without warning. If both flags are set, upload first tries to mere folders and only overwrites files. If ignore_hidden is set to true, hidden files are omitted in the upload. The start sequence of String or Char, with which a hidden file starts, can be specified by the hide_identifier. By default it is assumed that hidden files start with a dot (.).

Examples

sftp = SFTP.Client("sftp://test.rebex.net", "demo", "password")

upload(sftp, "data/test.csv", "/tmp") # upload data/test.csv to /tmp/test.csv

files=readdir()
upload.(sftp, files) # all files of the current directory are uploaded to the current directory on the server

upload(sftp, ignore_hidden=true) # the current folder is uploaded to the server without hidden objects

SFTP.upload_file — Function

upload_file(
    sftp::Client,
    src::AbstractString,
    dst::AbstractString,
    [files::Vector{<:AbstractString},
    conflicts::Vector{<:AbstractString},]
    force::Union{Nothing,Bool},
    ignore_hidden::Bool,
    hide_identifier::Union{AbstractString,Char}
)

Upload src to the dst path on the sftp client. If files and conflicts are given, all files within the src folder are uploaded to the dst folder and conflicts are handled according to the force flag. Otherwise. src must be an existing file that is uploaded to the dst folder.

If force is nothing, an error is thrown for existing files on dst, if force is true, existing files are overwritten, and if force is false, upload is skipped for existing files. Hidden files starting with the hide_identifier are ignored, if ignore_hidden is set to true.

SFTP.Client — Type

mutable struct SFTP.Client

SFTP manages the connection to the server and stores all relevant connection data.

Constructors

SFTP.Client(url::AbstractString, username::AbstractString, public_key_file::AbstractString, public_key_file::AbstractString; kwargs) -> SFTP.Client
SFTP.Client(url::AbstractString, username::AbstractString, password::AbstractString=""; kwargs) -> SFTP.Client

Construct an SFTP.Client from the url and either user information or public and private key file.

Arguments

url: The url to connect to, e.g., sftp://mysite.com

username/password: user credentials

public_key_file/public_key_file: authentication certificates

Keyword arguments

create_known_hosts_entry: Automatically create an entry in known hosts

disable_verify_peer: verify the authenticity of the peer's certificate

disable_verify_host: verify the host in the server's TLS certificate

verbose: display a lot of verbose curl information

Important notice

A username must be provided for both methods to work.

Before using the constructor method for certificate authentication, private and public key files must be created and stored in the ~/.ssh folder and on the server, e.g. ~/.ssh/idrsa and ~/.ssh/idrsa.pub. Additionally, the host must be added to the known_hosts file in the ~/.ssh folder.

The correct setup can be tested in the terminal with ssh myuser@mysitewhereIhaveACertificate.com.

Examples

sftp = SFTP.Client("sftp://mysitewhereIhaveACertificate.com", "myuser", "test.pub", "test.pem")
sftp = SFTP.Client("sftp://mysitewhereIhaveACertificate.com", "myuser")
sftp = SFTP.Client("sftp://test.rebex.net", "demo", "password")

SFTP.StatStruct — Type

struct SFTP.StatStruct

Hold information for file system objects on a Server.

Fields

desc::String: file or folder description/name

mode::UInt: file system object type (file, folder, etc.)

nlink::Int: number of hard links (contents)

uid::String: numeric user ID of the owner of the file/folder

uid::String: numeric group ID (gid) for the file/folder

size::Int64: file/folder size in Byte

mtime::Float64: modified time

Constructors

SFTP.StatStruct(stats::AbstractString) -> SFTP.StatStruct

Parse the stats string and return an SFTP.StatStruct.

The stats are of the format:

"d--x--x---  151 ftp      ftp          8192 Dec  2  2023 .."